+++ rmosh/src/vm.rs
@@ -5,11 +5,12 @@ use std::{
 };
 
 use crate::{
+    op::Op,
     compiler,
     equal::Equal,
     fasl::Fasl,
     gc::{Gc, GcRef},
-    objects::{Closure, Object, Op, Pair, Symbol, Vox},
+    objects::{Closure, Object, Pair, Symbol, Vox},
     procs::{self, default_free_vars},
 };
